% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_adder.R
\name{adder}
\alias{adder}
\title{ADD two numbers}
\usage{
adder(x, y)
}
\arguments{
\item{x}{a reel number}

\item{y}{another number}
}
\value{
returns the result of addition
}
\description{
This is a riveting description that describes what function does.
This is a link to \code{\link{hello}}
}
\examples{
adder(1,2)
}
\seealso{
\code{\link{hello}}
}
